Rename header files
